backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
yfinance==0.2.28
sqlalchemy==2.0.23
boto3==1.34.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
requests>=2.32.2
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.12.0
anthropic==0.18.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
mangum==0.17.0
aws-xray-sdk==2.12.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

